Changed to 16 bit

+++ Inc/mixer.h
@@ -1,8 +1,8 @@
 #include "main.h"
 
 extern void Mixer_AppendUSBBuffer(uint8_t * data, uint32_t size);
-extern void Mixer_AppendBTBuffer(uint32_t * data, uint32_t size);
-extern void Mixer_AppendWIFIBuffer(uint32_t * data, uint32_t size);
+extern void Mixer_AppendBTBuffer(uint16_t * data, uint32_t size);
+extern void Mixer_AppendWIFIBuffer(uint16_t * data, uint32_t size);
 extern void Mixer_TimerCallback(void);
 extern void Mixer_SyncUSB(void);
 extern void Mixer_SyncBT(void);
